requirements.txt
flask==3.0.0
requests==2.32.4
PyYAML==6.0.2
SQLAlchemy==2.0.38
psycopg2-binary==2.9.10
alembic==1.14.1
tzdata==2025.2
